% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_activity_frequency.R
\name{activity_frequency}
\alias{activity_frequency}
\title{Check activity frequencies}
\usage{
activity_frequency(activity_log, ..., details = TRUE,
  filter_condition = NULL)
}
\arguments{
\item{activity_log}{The activity log (renamed/formatted using functions rename_activity_log and convert_timestamp_format)}

\item{...}{Vectors of length 2 containing first the activity label followed by a numeric threshold value for that activity.}

\item{details}{Boolean indicating wheter details of the results need to be shown}

\item{filter_condition}{Condition that is used to extract a subset of the activity log prior to the application of the function}
}
\value{
An overview of frequencies of activities
}
\description{
Function that detects activity frequency anomalies per case
}
