% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_attribute_dependency_violations.R
\name{attribute_dependency}
\alias{attribute_dependency}
\title{Detect dependency violations between attributes}
\usage{
attribute_dependency(activity_log, condition_vector1, condition_vector2,
  details = TRUE, filter_condition = NULL)
}
\arguments{
\item{activity_log}{The activity log (renamed/formatted using functions rename_activity_log and convert_timestamp_format)}

\item{condition_vector1}{Vector of condition(s) which serve as an antecedent (if the condition(s) in condition_vector1 hold, then the condition(s) in condition_vector2 should also hold)}

\item{condition_vector2}{Vector of condition(s) which serve as a consequent (if the condition(s) in condition_vector1 hold, then the condition(s) in condition_vector2 should also hold)}

\item{details}{Boolean indicating wheter details of the results need to be shown}

\item{filter_condition}{Condition that is used to extract a subset of the activity log prior to the application of the function}
}
\value{
Information on the degree to which the specified dependencies are respected/violated.
}
\description{
Function detecting violations of dependencies between attributes (i.e. condition(s) that should hold when (an)other condition(s) hold(s))
}
