% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_attribute_range_violations.R
\name{attribute_range}
\alias{attribute_range}
\title{Detect attribute range violations}
\usage{
attribute_range(activity_log, column, domain_range,
  timestamp_format = "yyyy-mm-dd hh:mm:ss", details = TRUE,
  filter_condition = NULL)
}
\arguments{
\item{activity_log}{The activity log (renamed/formatted using functions rename_activity_log and convert_timestamp_format)}

\item{column}{Column in the activity log for which the domain range needs to be checked}

\item{domain_range}{Domain range for the specified column in parameter column}

\item{timestamp_format}{When the provided domain range represents timestamps, the format of the timestamps needs to be provided (either "yyyy-mm-dd hh:mm:ss" or "dd-mm-yyyy hh:mm:ss")}

\item{details}{Boolean indicating wheter details of the results need to be shown}

\item{filter_condition}{Condition that is used to extract a subset of the activity log prior to the application of the function}
}
\value{
Information on the presence of attribute range violations in the specified column
}
\description{
Function detecting violations of the attribute range, i.e. attribute values outside the range of tolerable values
}
