% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_colnames.R
\name{check_colnames}
\alias{check_colnames}
\title{Check log column names presence}
\usage{
check_colnames(log, ...)
}
\arguments{
\item{log}{The log whose columns names will be checked}

\item{...}{The labels to search for in the log columns}
}
\value{
A vector of labels that were not found in the log
}
\description{
Function that checks if the provided labels correspond to a column name in the log. FOR INTERNAL USE ONLY.
}
