% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_timestamp_format.R
\name{convert_timestamp_format}
\alias{convert_timestamp_format}
\title{Convert timestamp format}
\usage{
convert_timestamp_format(renamed_activity_log,
  timestamp_format = "yyyy-mm-dd hh:mm:ss")
}
\arguments{
\item{renamed_activity_log}{The activity log (renamed/formatted using functions rename_activity_log and convert_timestamp_format)}

\item{timestamp_format}{The format of the timestamps in the original dataset (either "yyyy-mm-dd hh:mm:ss" or "dd-mm-yyyy hh:mm:ss" or "yyyy/mm/dd hh:mm:ss" or "dd/mm/yyyy hh:mm:ss" or "yyyy-mm-dd" or "dd-mm-yyyy" or "yyyy/mm/dd" or "dd/mm/yyyy")}
}
\value{
Activity log with timestamps in converted timestamp format
}
\description{
Function converting the timestamps in the activity log to the appropriate format, required for the application of quality assessment functions
}
