% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daqapo.R
\docType{package}
\name{DaQAPO}
\alias{DaQAPO}
\alias{DaQAPO-package}
\title{DaQAPO - Data Quality Assessment for Process-oriented Data}
\description{
This package is designed to perform data quality assessment on process-oriented data.
}
