% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_activity_frequency_violations.R
\name{detect_activity_frequency_violations}
\alias{detect_activity_frequency_violations}
\alias{detect_activity_frequency_violations.activitylog}
\title{Check activity frequencies}
\usage{
detect_activity_frequency_violations(activity_log, ..., details,
  filter_condition)

\method{detect_activity_frequency_violations}{activitylog}(activity_log,
  ..., details = TRUE, filter_condition = NULL)
}
\arguments{
\item{activity_log}{The activity log}

\item{...}{Named vectors with name of the activity, and value of the threshold.}

\item{details}{Boolean indicating wheter details of the results need to be shown}

\item{filter_condition}{Condition that is used to extract a subset of the activity log prior to the application of the function}
}
\value{
An overview of frequencies of activities
}
\description{
Function that detects activity frequency anomalies per case
}
\section{Methods (by class)}{
\itemize{
\item \code{activitylog}: Detect activity frequency voilations in activity log
}}

