% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_attribute_dependency_violations.R
\name{detect_attribute_dependencies}
\alias{detect_attribute_dependencies}
\title{Detect dependency violations between attributes}
\usage{
detect_attribute_dependencies(activity_log, antecedent, consequence,
  details = TRUE, filter_condition = NULL, ...)
}
\arguments{
\item{activity_log}{The activity log (renamed/formatted using functions rename_activity_log and convert_timestamp_format)}

\item{antecedent}{(Vector of) condition(s) which serve as an antecedent (if the condition(s) in antecedent hold, then the condition(s) in consequence should also hold)}

\item{consequence}{(Vector of) condition(s) which serve as a consequent (if the condition(s) in antecedent hold, then the condition(s) in consequence should also hold)}

\item{details}{Boolean indicating wheter details of the results need to be shown}

\item{filter_condition}{Condition that is used to extract a subset of the activity log prior to the application of the function}
}
\value{
Information on the degree to which the specified dependencies are respected/violated.
}
\description{
Function detecting violations of dependencies between attributes (i.e. condition(s) that should hold when (an)other condition(s) hold(s))
}
