% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/detect_conditional_activity_presence_violations.R
\name{detect_conditional_activity_presence}
\alias{detect_conditional_activity_presence}
\title{Detect conditional activity presence violations}
\usage{
detect_conditional_activity_presence(activity_log, condition_vector,
  activity_vector, details = TRUE, filter_condition = NULL)
}
\arguments{
\item{activity_log}{The activity log (renamed/formatted using functions rename_activity_log and convert_timestamp_format)}

\item{condition_vector}{Vector of condition(s) which serve as an antecedent (if the condition(s) in condition_vector hold, then the activity/activities in activity_vector should be recorded)}

\item{activity_vector}{Vector of activity/activities which serve as a consequent (if the condition(s) in condition_vector hold, then the activity/activities in activity_vector should be recorded)}

\item{details}{Boolean indicating wheter details of the results need to be shown}

\item{filter_condition}{Condition that is used to extract a subset of the activity log prior to the application of the function}
}
\value{
Information on the degree to which the specified conditional activity presence is respected/violated.
}
\description{
Function detecting violations of conditional activity presence (i.e. an activity/activities that should be present when (a) particular condition(s) hold(s))
}
