% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_overlaps.R
\name{detect_overlaps}
\alias{detect_overlaps}
\title{Detect overlapping acitivity instances}
\usage{
detect_overlaps(
  activitylog,
  details = FALSE,
  level_of_aggregation = c("case", "resource"),
  filter_conditon
)
}
\arguments{
\item{activitylog}{The activity log}

\item{details}{Boolean indicating wheter details of the results need to be shown}

\item{level_of_aggregation}{Look for overlapping activity instances within a case or within a resource.}
}
\value{

}
\description{
Detect overlapping acitivity instances
}
