% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_character.R
\name{domain_categorical}
\alias{domain_categorical}
\title{Define allowable range of values}
\usage{
domain_categorical(allowed)
}
\arguments{
\item{allowed}{Allowed values of categorical column (character or factor)}
}
\value{
No return value, called for side effects
}
\description{
Define allowable range of values
}
\seealso{
\code{\link{detect_value_range_violations}}
}
