% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_numeric.R
\name{domain_numeric}
\alias{domain_numeric}
\title{Define allowable range of values}
\usage{
domain_numeric(from, to)
}
\arguments{
\item{from}{Minimum of allowed range}

\item{to}{Maximum of allowed range}
}
\value{
No return value, called for side effects
}
\description{
Define allowable range of values
}
\seealso{
\code{\link{detect_value_range_violations}}
}
