% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_time.R
\name{domain_time}
\alias{domain_time}
\title{Define allowable time range}
\usage{
domain_time(from, to, format = ymd_hms)
}
\arguments{
\item{from}{Start time interval}

\item{to}{End time interval}

\item{format}{Format of to and from (either ymd_hms, dmy_hms, ymd_hm, ymd, dmy, dmy, ...). Both from and to should have the same format.}
}
\value{
No return value, called for side effects
}
\description{
Define allowable time range
}
\seealso{
\code{\link{detect_value_range_violations}}
}
