% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_duration_outliers.R
\name{duration_outliers}
\alias{duration_outliers}
\title{Detect activity duration outliers}
\usage{
duration_outliers(activity_log, activity_considered, bound_sd = 3,
  lower_bound = NA, upper_bound = NA, details = TRUE,
  filter_condition = NULL)
}
\arguments{
\item{activity_log}{The activity log (renamed/formatted using functions rename_activity_log and convert_timestamp_format)}

\item{activity_considered}{Activity name for which duration outliers should be detected}

\item{bound_sd}{Number of standard deviations from the mean duration which is used to define an outlier in the absence of lower_bound and upper_bound (default value of 3 is used)}

\item{lower_bound}{Lower bound for activity duration used during outlier detection (expressed in minutes)}

\item{upper_bound}{Upper bound for activity duration used during outlier detection (expressed in minutes)}

\item{details}{Boolean indicating wheter details of the results need to be shown}

\item{filter_condition}{Condition that is used to extract a subset of the activity log prior to the application of the function}
}
\value{
Information on the presence of activity duration outliers
}
\description{
Function detecting duration outliers for a particular activity
}
