% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_duration_outliers.R
\name{duration_within}
\alias{duration_within}
\title{Detect activity duration outliers}
\usage{
duration_within(bound_sd = 3, lower_bound = NA, upper_bound = NA)
}
\arguments{
\item{bound_sd}{Number of standard deviations from the mean duration which is used to define an outlier in the absence of lower_bound and upper_bound (default value of 3 is used)}

\item{lower_bound}{Lower bound for activity duration used during outlier detection (expressed in minutes)}

\item{upper_bound}{Upper bound for activity duration used during outlier detection (expressed in minutes)}

\item{activity_considered}{Activity name for which duration outliers should be detected}
}
\value{
Information on the presence of activity duration outliers
}
\description{
Function detecting duration outliers for a particular activity
}
