% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_to_activities.R
\name{events_to_activities}
\alias{events_to_activities}
\alias{events_to_activities.eventlog}
\title{Events to activities}
\usage{
events_to_activities(eventlog)

\method{events_to_activities}{eventlog}(eventlog)
}
\arguments{
\item{eventlog}{The event log to be converted. An object of class
\code{eventlog}.}
}
\description{
Events to activities
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Create activity log from event log
}}

