% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_anomalies.R
\name{filter_anomalies}
\alias{filter_anomalies}
\title{Filter anomalies from the activity log}
\usage{
filter_anomalies(activity_log, anomaly_log)
}
\arguments{
\item{activity_log}{The activity log (renamed/formatted using functions rename_activity_log and convert_timestamp_format)}

\item{anomaly_log}{The anomaly log generated from the different DAQAPO tests}
}
\value{
activitylog in which the anomaly rows are filtered out
}
\description{
Function that filters detected anomalies from the activity log
}
