% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hospital_actlog}
\alias{hospital_actlog}
\title{An activity log of 20 patients in a hospital (activity log object)}
\format{
An activity log with 53 rows and 7 variables:
\describe{
  \item{patient_visit_nr}{the patient's identifier}
  \item{activity}{the executed activity}
  \item{originator}{the resource performing the activity execution}
  \item{start}{the timestamp at which the activity was started}
  \item{complete}{the timestamp at which the activity was completed}
  \item{triagecode}{a case attribute describing the triage code}
  \item{specialization}{a case attribute describing the specialization}
}
}
\source{
An illustrative example developed in-house for demonstrational purposes.
}
\usage{
hospital_actlog
}
\description{
A dataset containing the logged activities in an illustrative hospital process.
20 patients are described in the log.
Process activities include Registration, Triage, Clinical exam, Treatment and Treatment evaluation.
}
\keyword{datasets}
