% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_incorrect_activity_names.R
\name{incorrect_activity_names}
\alias{incorrect_activity_names}
\title{Detect incorrect activity names}
\usage{
incorrect_activity_names(activity_log, details = TRUE,
  filter_condition = NULL)
}
\arguments{
\item{activity_log}{The activity log (renamed/formatted using functions rename_activity_log and convert_timestamp_format)}

\item{details}{Boolean indicating wheter details of the results need to be shown}

\item{filter_condition}{Condition that is used to extract a subset of the activity log prior to the application of the function}
}
\value{
Information on the incorrect activities in the log
}
\description{
Function returning the incorrect activity labels in the log as indicated by the user. If details are requested, the entire activity log's rows containing incorrect activities are returned.
}
