% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_activity_log.R
\name{read_activity_log}
\alias{read_activity_log}
\title{Read activity log dataset from csv}
\usage{
read_activity_log(file_name, case_id_label = NULL,
  activity_label = NULL, resource_label = NULL, start_label = NULL,
  complete_label = NULL, case_attributes = NULL,
  timestamp_format = NULL, sep = ",", empty_cells_are_NA = TRUE,
  zero_values_are_NA = FALSE)
}
\arguments{
\item{file_name}{The file name of the input csv dataset}

\item{case_id_label}{The name of the column in the original dataset reflecting the case identifier}

\item{activity_label}{The name of the column in the original dataset reflecting the activity name}

\item{resource_label}{The name of the column in the original dataset reflecting the resource identifier}

\item{start_label}{The name of the column in the original dataset reflecting the activity start times}

\item{complete_label}{The name of the column in the original dataset reflecting the activity completion times}

\item{case_attributes}{A vector containing the name(s) of the column(s) in the original dataset reflecting case attributes}

\item{timestamp_format}{The format of the timestamps in the original dataset (either "yyyy-mm-dd hh:mm:ss" or "dd-mm-yyyy hh:mm:ss")}

\item{sep}{The character that separates columns in csv file formats}

\item{empty_cells_are_NA}{Boolean indicating whether empty cells should be considered as NA}

\item{zero_values_are_NA}{Boolean indicating whether zero values (i.e. 0) should be considered as NA}
}
\value{
Input csv dataset with appropriate formatting of NA values
}
\description{
Function which reads input csv dataset. Depending on the specification of the function's parameters, it also replaces particular cell values by NA.
}
