% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_event_log.R
\name{read_event_log}
\alias{read_event_log}
\title{Read an event log dataset}
\usage{
read_event_log(file_name, case_id_label = NULL, activity_label = NULL,
  resource_label = NULL, event_lifecycle_label = NULL,
  timestamps_label = NULL, case_attributes = NULL,
  timestamp_format = NULL, event_matching_label = NULL, sep = ",",
  empty_cells_are_NA = TRUE, zero_values_are_NA = FALSE)
}
\arguments{
\item{file_name}{The file name of the input dataset. Can be csv, xes or eventlog format}

\item{case_id_label}{The name of the column in the original dataset reflecting the case identifier}

\item{activity_label}{The name of the column in the original dataset reflecting the activity name}

\item{resource_label}{The name of the column in the original dataset reflecting the resource identifier}

\item{event_lifecycle_label}{The name of the column in the original dataset reflecting the activity lifecycle state}

\item{timestamps_label}{The name of the column in the original dataset reflecting the activity times}

\item{case_attributes}{A vector containing the name(s) of the column(s) in the original dataset reflecting case attributes}

\item{timestamp_format}{The format of the timestamps in the original dataset (either "yyyy-mm-dd hh:mm:ss" or "dd-mm-yyyy hh:mm:ss")}

\item{event_matching_label}{The name of the column in the original dataset reflecting the relation between start and end events per case per activity}

\item{sep}{The character that separates columns in csv file formats}

\item{empty_cells_are_NA}{Boolean indicating whether empty cells should be considered as NA}

\item{zero_values_are_NA}{Boolean indicating whether zero values (i.e. 0) should be considered as NA}
}
\value{
An activity log with proper NA settings, data restructuring and type conversions.
}
\description{
Function which reads event log datasets. Depending on the specification of the function's parameters, it also replaces particular cell values by NA.
}
