% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_input_csv.R
\name{read_input_csv}
\alias{read_input_csv}
\title{Read input csv dataset}
\usage{
read_input_csv(file_name, empty_cells_are_NA = TRUE,
  zero_values_are_NA = FALSE)
}
\arguments{
\item{file_name}{The file name of the input csv dataset}

\item{empty_cells_are_NA}{Boolean indicating whether empty cells should be considered as NA}

\item{zero_values_are_NA}{Boolean indicating whether zero values (i.e. 0) should be considered as NA}
}
\value{
Input csv dataset with appropriate formatting of NA values
}
\description{
Function which reads input csv dataset. Depending on the specification of the function's parameters, it also replaces particular cell values by NA.
}
