% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_missing_related_activities.R
\name{related_activities}
\alias{related_activities}
\title{Detect missing related activities}
\usage{
related_activities(activity_log, activity1, activity2, details = TRUE,
  filter_condition = NULL)
}
\arguments{
\item{activity_log}{The activity log (renamed/formatted using functions rename_activity_log and convert_timestamp_format)}

\item{activity1}{Activity name of the activity that acts as a an antecedent (if activity1 occurs, then activity2 should also occur)}

\item{activity2}{Activity name of the activity that acts as a an consequent (if activity1 occurs, then activity2 should also occur)}

\item{details}{Boolean indicating wheter details of the results need to be shown}

\item{filter_condition}{Condition that is used to extract a subset of the activity log prior to the application of the function}
}
\value{
Information on the presence/absence of related activities.
}
\description{
Function detecting missing related activity registration, i.e. detecting activities that should be registered for a case because another activity is registered for that case
}
