% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_activity_log.R
\name{rename_activity_log}
\alias{rename_activity_log}
\title{Rename columns in an activity log}
\usage{
rename_activity_log(activity_log, case_id_label, activity_label,
  resource_label, start_time_label, completion_time_label,
  case_attributes = NULL)
}
\arguments{
\item{activity_log}{The activity log}

\item{case_id_label}{The name of the column in the original dataset reflecting the case identifier}

\item{activity_label}{The name of the column in the original dataset reflecting the activity name}

\item{resource_label}{The name of the column in the original dataset reflecting the resource identifier}

\item{start_time_label}{The name of the column in the original dataset reflecting the activity start times}

\item{completion_time_label}{The name of the column in the original dataset reflecting the activity completion times}

\item{case_attributes}{A vector containing the name(s) of the column(s) in the original dataset reflecting case attributes}
}
\value{
Activity log with renamed columns
}
\description{
Function renaming the columns in the activity log to the appropriate names, required for the application of quality assessment functions
}
