% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_resource_inconsistencies.R
\name{resource_inconsistencies}
\alias{resource_inconsistencies}
\title{Detect resource inconsistencies}
\usage{
resource_inconsistencies(event_log, case_id_label, activity_label,
  resource_label, event_lifecycle_label, timestamps_label,
  event_matching_label = NULL, details = TRUE,
  filter_condition = NULL, detect_only = TRUE)
}
\arguments{
\item{event_log}{The data frame containing the event log data}

\item{case_id_label}{The name of the column in the original dataset reflecting the case identifier}

\item{activity_label}{The name of the column in the original dataset reflecting the activity name}

\item{resource_label}{The name of the column in the original dataset reflecting the resource identifier}

\item{event_lifecycle_label}{The name of the column in the original dataset reflecting the activity lifecycle state}

\item{timestamps_label}{The name of the column in the original dataset reflecting the activity times}

\item{event_matching_label}{The name of the column in the original dataset reflecting the relation between start and end events per case per activity}

\item{details}{Boolean indicating wheter details of the results need to be shown}

\item{filter_condition}{Condition that is used to extract a subset of the activity log prior to the application of the function}

\item{detect_only}{Boolean indiciating what results to return: if TRUE, immediatly return the detection of inconsistencies. If FALSE, return a new log with resolved issues if the users requests this correction}
}
\value{
an overview containing the activity executions that do not have matching resources for their start and complete events OR the new event log with all resource inconsistencies corrected.
}
\description{
Function to detect inconsistencies in resource information between related start and complete events.
}
