% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restructure_to_activity_log.R
\name{restructure_to_activity_log}
\alias{restructure_to_activity_log}
\title{Restructure an event log dataset to activity log format}
\usage{
restructure_to_activity_log(event_log, case_id_label, activity_label,
  resource_label, event_lifecycle_label, timestamps_label,
  case_attributes = NULL, event_matching_label = NULL,
  timestamp_format = "yyyy-mm-dd hh:mm:ss")
}
\arguments{
\item{event_log}{The data frame containing the event log data}

\item{case_id_label}{The name of the column in the original dataset reflecting the case identifier}

\item{activity_label}{The name of the column in the original dataset reflecting the activity name}

\item{resource_label}{The name of the column in the original dataset reflecting the resource identifier}

\item{event_lifecycle_label}{The name of the column in the original dataset reflecting the activity lifecycle state}

\item{timestamps_label}{The name of the column in the original dataset reflecting the activity times}

\item{case_attributes}{A vector containing the name(s) of the column(s) in the original dataset reflecting case attributes}

\item{event_matching_label}{The name of the column in the original dataset reflecting the relation between start and end events per case per activity. If NULL, a basic heuristic will be applied that links every first available start event with the first available complete event}

\item{timestamp_format}{The format of the timestamps in the original dataset (either "yyyy-mm-dd hh:mm:ss" or "dd-mm-yyyy hh:mm:ss")}
}
\value{
A data frame in the format of an activity log.
}
\description{
Function which takes an event log as input and transforms it into an activity log. If event_matching_label, the name of the column that contains information which start and complete events belong together, is NULL, a basic heuristic of "for every start event per case per activity, take the first available complete event." will be applied.
}
