% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_similar_labels.R
\name{similar_labels}
\alias{similar_labels}
\title{Search for similar labels in a column}
\usage{
similar_labels(activity_log, column_label, max_edit_distance = 3,
  show_NA = FALSE, ignore_capitals = FALSE)
}
\arguments{
\item{activity_log}{The activity log (renamed/formatted using functions rename_activity_log and convert_timestamp_format)}

\item{column_label}{The name of the column in which to search for spelling mistakes}

\item{max_edit_distance}{The maximum number of insertions, deletions and substitutions that are allowed to be executed in order for two strings to be considered similar.}

\item{show_NA}{A boolean indicating if labels that do not show similarities with others should be shown in the output}

\item{ignore_capitals}{A boolean indicating if capitalization should be included or excluded when calculating the edit distance between two strings}
}
\value{
A data frame containing, for each label in the column, an overview of similar labels.
}
\description{
Function that tries to detect spelling mistakes in a given activity log column
}
