% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_time_anomalies.R
\name{time_anomalies}
\alias{time_anomalies}
\title{Detect time anomalies}
\usage{
time_anomalies(activity_log, anomaly_type = "both", details = TRUE,
  filter_condition = NULL)
}
\arguments{
\item{activity_log}{The activity log (renamed/formatted using functions rename_activity_log and convert_timestamp_format)}

\item{anomaly_type}{Type of anomalies that need to be detected (either "negative", "zero" or "both")}

\item{details}{Boolean indicating wheter details of the results need to be shown}

\item{filter_condition}{Condition that is used to extract a subset of the activity log prior to the application of the function}
}
\value{
Information on the presence of time anomalies of the specified anomaly type
}
\description{
Function detecting time anomalies, which can refer to activities with negative or zero duration
}
