% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_unique_values.R
\name{unique_values}
\alias{unique_values}
\title{Search for unique values / distinct combinations}
\usage{
unique_values(activity_log, ..., filter_condition = NULL)
}
\arguments{
\item{activity_log}{The activity log (renamed/formatted using functions rename_activity_log and convert_timestamp_format)}

\item{...}{The names of columns in the activity log for which you want to show the different combinations found in the log. If only one column is provided, this results in a list of unique values in that column.}

\item{filter_condition}{Condition that is used to extract a subset of the activity log prior to the application of the function}
}
\value{
A data frame containing the unique (distinct) values (combinations) in the requested column(s).
}
\description{
Function that lists all distinct combinations of the given columns in the activity log
}
