% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_activity_order_violations.R
\name{detect_activity_order_violations}
\alias{detect_activity_order_violations}
\alias{detect_activity_order_violations.activitylog}
\title{Detect activity order violations}
\usage{
detect_activity_order_violations(activity_log, activity_order, timestamp,
  details, filter_condition)

\method{detect_activity_order_violations}{activitylog}(activity_log,
  activity_order, timestamp = c("both", "start", "complete"),
  details = TRUE, filter_condition = NULL)
}
\arguments{
\item{activity_log}{The activity log}

\item{activity_order}{Vector expressing the activity order that needs to be checked (using activity names)}

\item{timestamp}{Type of timestamp that needs to be taken into account in the analysis (either "start", "complete" or "both)}

\item{details}{Boolean indicating wheter details of the results need to be shown}

\item{filter_condition}{Condition that is used to extract a subset of the activity log prior to the application of the function}
}
\value{
Information on the degree to which the specified activity order is respected/violated.
}
\description{
Function detecting violations in activity order
}
\section{Methods (by class)}{
\itemize{
\item \code{activitylog}: Detect activity order_violations in activity log.
}}

