% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_case_id_sequence.R
\name{detect_case_id_sequence_gaps}
\alias{detect_case_id_sequence_gaps}
\title{Detect gaps in case_id}
\usage{
detect_case_id_sequence_gaps(activity_log, details = TRUE,
  filter_condition = NULL)
}
\arguments{
\item{activity_log}{The activity log (renamed/formatted using functions rename_activity_log and convert_timestamp_format)}

\item{details}{Boolean indicating wheter details of the results need to be shown}

\item{filter_condition}{Condition that is used to extract a subset of the activity log prior to the application of the function}
}
\value{
The case IDs that are missing in the sequence.
}
\description{
Function detecting gaps in the sequence of case identifiers
}
