% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration_within.R
\name{duration_within}
\alias{duration_within}
\title{Define bounds for activity duration}
\usage{
duration_within(bound_sd = 3, lower_bound = NA, upper_bound = NA)
}
\arguments{
\item{bound_sd}{Number of standard deviations from the mean duration which is used to define an outlier in the absence of lower_bound and upper_bound (default value of 3 is used)}

\item{lower_bound}{Lower bound for activity duration used during outlier detection (expressed in minutes). This means disregarding the sd and bound_sd for lower bound}

\item{upper_bound}{Upper bound for activity duration used during outlier detection (expressed in minutes). This means disregarding the sd and bound_sd for upper bound}
}
\value{
No return value, called for side effects
}
\description{
Funtion to define bounds on the duration of an activity during detection of duration outliers.
}
\seealso{
\code{\link{detect_duration_outliers}}
}
