% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.R
\name{fix}
\alias{fix}
\title{Fix problems}
\usage{
fix(detected_problems, ...)
}
\arguments{
\item{detected_problems}{Output of a detect_ function. Currently supported: detect_resource_inconsistencies.}

\item{...}{Additionals parameters, depending on type of anomalies to fix.}
}
\description{
Fix problems
}
